Added basic physics engine
I would like this engine to be general enough to allow the 2d
platforming required in homework 1 section 1, as well as a specific 2d
top-down game I'd like to develop that has many moving physics bits, so
I will be using a physics library later. For assignment 1 I decided to
make calls to some physics engine interface that I would implement
temporarily with something simple to get the jumping and collision
detection done.